need to tweak logic for lookups that hit on the current line
